Update `3-logreg-nb-imdb.ipynb` notebook
In this updated version of the `3-logreg-nb-imdb.ipynb` notebook, I've resolved all the outstanding issues, as well as done a fair bit of refactoring, annotation, and development along the lines of the original lesson. New material includes more on `Compressed Sparse Row` format (section 6), an extensive discussion of the `Naive Bayes Classifier` (sections 8 and 9), a brief introduction to `Fastai's API for text data sets` (section 11), a summary table (section 17) , as well as extensive explanatory notes throughout the notebook.